requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
distro==1.9.0
Django==5.2.3
django-environ==0.12.0
djangorestframework==3.16.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
faiss-cpu==1.11.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
numpy==2.2.6
openai==1.93.0
outcome==1.3.0.post0
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
selenium==4.34.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
